// packages/core/src/generated/message.ts
export const MSG = {
  GENERIC: 0,
  BIRTH: 1,
  HIT: 2,
  MISS: 3,
  FLEE: 4,
  DROP: 5,
  KILL: 6,
  LEVEL: 7,
  DEATH: 8,
  STUDY: 9,
  TELEPORT: 10,
  SHOOT: 11,
  QUAFF: 12,
  ZAP_ROD: 13,
  WALK: 14,
  TPOTHER: 15,
  HITWALL: 16,
  EAT: 17,
  STORE1: 18,
  STORE2: 19,
  STORE3: 20,
  STORE4: 21,
  DIG: 22,
  OPENDOOR: 23,
  SHUTDOOR: 24,
  TPLEVEL: 25,
  BELL: 26,
  NOTHING_TO_OPEN: 27,
  LOCKPICK_FAIL: 28,
  STAIRS_DOWN: 29,
  HITPOINT_WARN: 30,
  ACT_ARTIFACT: 31,
  USE_STAFF: 32,
  DESTROY: 33,
  MON_HIT: 34,
  MON_TOUCH: 35,
  MON_PUNCH: 36,
  MON_KICK: 37,
  MON_CLAW: 38,
  MON_BITE: 39,
  MON_STING: 40,
  MON_BUTT: 41,
  MON_CRUSH: 42,
  MON_ENGULF: 43,
  MON_CRAWL: 44,
  MON_DROOL: 45,
  MON_SPIT: 46,
  MON_GAZE: 47,
  MON_WAIL: 48,
  MON_SPORE: 49,
  MON_BEG: 50,
  MON_INSULT: 51,
  MON_MOAN: 52,
  RECOVER: 53,
  BLIND: 54,
  CONFUSED: 55,
  POISONED: 56,
  AFRAID: 57,
  PARALYZED: 58,
  DRUGGED: 59,
  SPEED: 60,
  SLOW: 61,
  SHIELD: 62,
  BLESSED: 63,
  HERO: 64,
  BERSERK: 65,
  BOLD: 66,
  PROT_EVIL: 67,
  INVULN: 68,
  SEE_INVIS: 69,
  INFRARED: 70,
  RES_ACID: 71,
  RES_ELEC: 72,
  RES_FIRE: 73,
  RES_COLD: 74,
  RES_POIS: 75,
  STUN: 76,
  CUT: 77,
  STAIRS_UP: 78,
  STORE_ENTER: 79,
  STORE_LEAVE: 80,
  STORE_HOME: 81,
  MONEY1: 82,
  MONEY2: 83,
  MONEY3: 84,
  SHOOT_HIT: 85,
  STORE5: 86,
  LOCKPICK: 87,
  DISARM: 88,
  IDENT_BAD: 89,
  IDENT_EGO: 90,
  IDENT_ART: 91,
  BR_ELEMENTS: 92,
  BR_FROST: 93,
  BR_ELEC: 94,
  BR_ACID: 95,
  BR_GAS: 96,
  BR_FIRE: 97,
  BR_DISEN: 98,
  BR_CHAOS: 99,
  BR_SHARDS: 100,
  BR_SOUND: 101,
  BR_LIGHT: 102,
  BR_DARK: 103,
  BR_NETHER: 104,
  BR_NEXUS: 105,
  BR_TIME: 106,
  BR_INERTIA: 107,
  BR_GRAVITY: 108,
  BR_PLASMA: 109,
  BR_FORCE: 110,
  SUM_MONSTER: 111,
  SUM_AINU: 112,
  SUM_UNDEAD: 113,
  SUM_ANIMAL: 114,
  SUM_SPIDER: 115,
  SUM_HOUND: 116,
  SUM_HYDRA: 117,
  SUM_DEMON: 118,
  SUM_DRAGON: 119,
  SUM_HI_UNDEAD: 120,
  SUM_HI_DRAGON: 121,
  SUM_HI_DEMON: 122,
  SUM_WRAITH: 123,
  SUM_UNIQUE: 124,
  WIELD: 125,
  QUIVER: 126,
  CURSED: 127,
  RUNE: 128,
  HUNGRY: 129,
  NOTICE: 130,
  AMBIENT_DAY: 131,
  AMBIENT_NITE: 132,
  AMBIENT_DNG1: 133,
  AMBIENT_DNG2: 134,
  AMBIENT_DNG3: 135,
  AMBIENT_DNG4: 136,
  AMBIENT_DNG5: 137,
  CREATE_TRAP: 138,
  SHRIEK: 139,
  CAST_FEAR: 140,
  HIT_GOOD: 141,
  HIT_GREAT: 142,
  HIT_SUPERB: 143,
  HIT_HI_GREAT: 144,
  HIT_HI_SUPERB: 145,
  SPELL: 146,
  PRAYER: 147,
  KILL_UNIQUE: 148,
  KILL_KING: 149,
  DRAIN_STAT: 150,
  MULTIPLY: 151,
  SCRAMBLE: 152,
  MAX: 153,
} as const;